Move remote keys to new folder to keep separate.
